Title: 토끼와 경주, Time: 687ms, Memory: 25MB, Status: Passed - CodeTree

+++ 240916/토끼와 경주/rabit-and-race.java
@@ -71,7 +71,7 @@ public class Main {
         rabbits.add(rabbit);
     }
 
-    public static int total;
+    public static long total;
     // 200 경주 진행
     public static void play(int K, int S) {
         Set<Rabbit> choosed = new HashSet<>();
